Added new method to prevent generating overlapping spheres during random gen
This is implemented using new methods in myVector (distance function) and in sphere/plane/object as well as scene. To determine overlap we use distance from origins compared to sum of both radius.

The function getSpecialVal() in the object and inherited classes will likely be refactored to return a vector in the future in order to allow for more modularity.